api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Thimm <thimm@farmonapp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
requests = "^2.31.0"
pandas = "^2.2.0"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
pyarrow = "^15.0.0"
pystac = "^1.9.0"
python-dotenv = "^1.0.1"
pip = "^23.3.2"
pymongo = "^4.6.1"
motor = "^3.3.2"
shapely = "^2.0.2"
geopandas = "^0.14.2"
beanie = "^1.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ipykernel = "^6.29.0"
mongomock-motor = "^0.0.21"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
